apps/api/requirements.txt
# Web
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Cache / Queue
redis[asyncio]==5.1.1
celery[redis]==5.4.0

# HTTP
httpx==0.27.2

# Auth / Security
python-jose[cryptography]==3.3.0
cryptography==43.0.3
passlib==1.7.4

# Config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Payment
stripe==11.1.0

# QR code / Images
qrcode==7.4.2
Pillow==10.4.0
